[add] 이메일, 비밀번호 폼 정규표현식과 옵션 파일로 분리
