+++ libs/components/src/lib/ResetPasswordForm.tsx
@@ -17,6 +17,7 @@ import {
   useMailVerificationMutation,
   useCountdown,
 } from '@project-lc/hooks';
+import { emailCodeRegisterOptions, emailRegisterOptions } from '@project-lc/shared-types';
 import { useRouter } from 'next/router';
 import { useCallback, useState } from 'react';
 import { useForm } from 'react-hook-form';
@@ -181,13 +182,7 @@ export function ResetPasswordForm(): JSX.Element {
                 placeholder="minsu@example.com"
                 autoComplete="off"
                 autoFocus
-                {...register('email', {
-                  required: '이메일을 작성해주세요.',
-                  pattern: {
-                    value: /^[\w]+@[\w]+\.[\w][\w]+$/,
-                    message: '이메일 형식이 올바르지 않습니다.',
-                  },
-                })}
+                {...register('email', { ...emailRegisterOptions })}
               />
               <FormErrorMessage>{errors.email && errors.email.message}</FormErrorMessage>
             </FormControl>
@@ -216,17 +211,7 @@ export function ResetPasswordForm(): JSX.Element {
                 placeholder="이메일 인증코드"
                 autoComplete="off"
                 autoFocus
-                {...register('code', {
-                  required: '인증 코드를 입력해주세요.',
-                  minLength: {
-                    value: 6,
-                    message: '인증코드는 6자 입니다.',
-                  },
-                  maxLength: {
-                    value: 6,
-                    message: '인증코드는 6자 입니다.',
-                  },
-                })}
+                {...register('code', { ...emailCodeRegisterOptions })}
               />
               <FormErrorMessage>{errors.code && errors.code.message}</FormErrorMessage>
             </FormControl>
